Data persistance fixed, working on view output and updating with database changes

+++ MobileDev/3200AndroidStarterApp-main/API/src/main/java/com/connorosbornefitnesstrackerproject1/api/models/Goal.java
@@ -8,8 +8,8 @@ import com.google.firebase.database.IgnoreExtraProperties;
 @IgnoreExtraProperties
 public class Goal {
     public String task;
-    public int amount;
-    public int progress;
+    public long amount;
+    public long progress;
 
     @Exclude
     public String id;
